Update STAMPTEst.ioc
As I suggested: we will face a major problem during interrupt handling, we need to lower the scheduling and waiting intervall. Or we need to overthink our software design.

I suggest, that we will handle ADC interrupts in ISR and so thread related work there. This may will extend the interrupt service routine. We may need to implement our own system to without scheduling and threads just to reach the goal.